Feature CFL adaptation initial iteration (#2362)
* Feature to define initial iteration to apply CFL adaption.

A spark initialization was not possible with CFL adaptation, as the CFL number would increase too far before the spark initiates the combustion. This led to immediate flashback or blow-off. An additional parameter in the CFL adaptation defines the iteration, at which the CFL adaption should start. This allows the combustion to stabilize before the CFL number increases.

Co-authored-by: Nijso <nijso@hotmail.com>
Co-authored-by: TobiKattmann <31306376+TobiKattmann@users.noreply.github.com>
Co-authored-by: Nijso <bigfootedrockmidget@hotmail.com>
Co-authored-by: Pedro Gomes <38071223+pcarruscag@users.noreply.github.com>